requirements.txt
langchain
langchain-openai
langchain-groq
langchain-core
langchain-community
langchain-text-splitters
langchain_chroma
langchain_huggingface
langchain-ollama
sentence-transformers

python-dotenv

pypdf
bs4
arxiv
pymupdf
chromadb
wikipedia
faiss-cpu
duckdb
pandas
openai
ipykernel
streamlit